Added Head Model and adjusted URLs

+++ webapp/models.py
@@ -3,8 +3,8 @@ from django.db import models
 # Create your models here.
 class Head(models.Model):
     animal = models.CharField(max_length=64)
-    price_per_kilo = models.IntegerField()
-    weight = models.IntegerField()
+    price_per_kilo = models.DecimalField(max_digits=10, decimal_places=2)
+    weight = models.FloatField()
     age = models.IntegerField()
     ready_for_sale = models.BooleanField(default=False)
